requirements.txt
blinker==1.6.2
Bottleneck==1.3.7
click==8.1.7
emoji==2.12.1
Flask==3.0.3
greenlet==3.0.1
itsdangerous==2.2.0
jaro-winkler==2.0.3
Jinja2==3.1.4
joblib==1.4.0
MarkupSafe==2.1.3
mysql-connector-python==8.3.0
numexpr==2.8.7
numpy==1.26.4
pandas==2.2.1
pip==24.0
pyjarowinkler==1.8
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.11.4
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.25
threadpoolctl==2.2.0
typing_extensions==4.11.0
tzdata==2023.3
Unidecode==1.2.0
Werkzeug==3.0.3
wheel==0.43.0
